Move database utils into shared
