+++ enterprise/frontend/src/metabase-enterprise/writable_connection/components/WritableConnectionInfoSection/WritableConnectionInfoSection.tsx
@@ -6,12 +6,12 @@ import {
   DatabaseInfoSection,
   DatabaseInfoSectionDivider,
 } from "metabase/admin/databases/components/DatabaseInfoSection";
+import { useUpdateDatabaseMutation } from "metabase/api";
+import { useConfirmation } from "metabase/common/hooks/use-confirmation";
 import {
   hasDbRoutingEnabled,
   isDbModifiable,
-} from "metabase/admin/databases/utils";
-import { useUpdateDatabaseMutation } from "metabase/api";
-import { useConfirmation } from "metabase/common/hooks/use-confirmation";
+} from "metabase/common/utils/database";
 import type { WritableConnectionInfoSectionProps } from "metabase/plugins/oss/writable-connection";
 import { Alert, Button, Group, Icon } from "metabase/ui";
 import * as Urls from "metabase/utils/urls";
