+++ frontend/src/metabase/detail-view/components/DetailViewSidesheet/utils.ts
@@ -4,7 +4,7 @@ import {
   isImplicitDeleteAction,
   isImplicitUpdateAction,
 } from "metabase/actions/utils";
-import { hasActionsEnabled } from "metabase/admin/databases/utils";
+import { hasActionsEnabled } from "metabase/common/utils/database";
 import { extractRemappedColumns } from "metabase/visualizations";
 import { getQuestionIdFromVirtualTableId } from "metabase-lib/v1/metadata/utils/saved-questions";
 import { findColumnIndexesForColumnSettings } from "metabase-lib/v1/queries/utils/dataset";
