+++ frontend/src/metabase/transforms/utils.ts
@@ -1,8 +1,8 @@
 import { t } from "ttag";
 import _ from "underscore";
 
-import { hasFeature } from "metabase/admin/databases/utils";
 import type { OmniPickerCollectionItem } from "metabase/common/components/Pickers/EntityPicker/types";
+import { hasFeature } from "metabase/common/utils/database";
 import { parseTimestamp } from "metabase/utils/time-dayjs";
 import * as Lib from "metabase-lib";
 import type Metadata from "metabase-lib/v1/metadata/Metadata";
